timestamp_extractor/requirements.txt
kafka-python==2.0.2
pytesseract==0.3.7
Pillow==8.1.2